requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
pg8000
python-dotenv
gunicorn
requests
requests-oauthlib
PyJWT